+++ Project 2/Server/deleteLocationByID.php
@@ -13,18 +13,41 @@
 	// SQL statement accepts parameters and so is prepared to avoid SQL injection.
 	// $_REQUEST used for development / debugging. Remember to change to $_POST for production
 
-	$placeholders = implode(',', array_fill(0, count($_REQUEST['locations']), '?'));
-
-	$query = $conn->prepare("DELETE FROM location WHERE id IN ($placeholders)");
-
-	$types = str_repeat('i', count($_REQUEST['locations']));
-	$query->bind_param($types, ...$_REQUEST['locations']);
-
+	$locations = "('" . implode("','", $_REQUEST['locations']) . "')";
+	$query = $conn->prepare("SELECT * FROM department as d INNER JOIN location as l ON l.id = d.locationID WHERE l.id IN $locations ");
 	$query->execute();
-	
-	if (false === $query) {
-		$output['status']['code'] = "400";
-		$output['status']['name'] = "executed";
+	$result = $query->get_result();
+
+	if($result->num_rows == 0) {
+		$placeholders = implode(',', array_fill(0, count($_REQUEST['locations']), '?'));
+		$query = $conn->prepare("DELETE FROM location WHERE id IN ($placeholders)");
+
+		$types = str_repeat('i', count($_REQUEST['locations']));
+		$query->bind_param($types, ...$_REQUEST['locations']);
+		$query->execute();
+		
+		if (false === $query) {
+			$output['status']['code'] = "400";
+			$output['status']['name'] = "executed";
+			$output['status']['description'] = "query failed";	
+			$output['data'] = [];
+
+			mysqli_close($conn);
+			echo json_encode($output); 
+			exit;
+		}
+
+		$output['status']['code'] = "200";
+		$output['status']['name'] = "ok";
+		$output['status']['description'] = "success";
+		$output['status']['returnedIn'] = (microtime(true) - $executionStartTime) / 1000 . " ms";
+		$output['data'] = [];
+		
+		mysqli_close($conn);
+		echo json_encode($output);
+	}else {
+		$output['status']['code'] = "409";
+		$output['status']['name'] = "conflict";
 		$output['status']['description'] = "query failed";	
 		$output['data'] = [];
 
@@ -32,15 +55,4 @@
 		echo json_encode($output); 
 		exit;
 	}
-
-	$output['status']['code'] = "200";
-	$output['status']['name'] = "ok";
-	$output['status']['description'] = "success";
-	$output['status']['returnedIn'] = (microtime(true) - $executionStartTime) / 1000 . " ms";
-	$output['data'] = [];
-	
-	mysqli_close($conn);
-
-	echo json_encode($output); 
-
 ?>
